Remove top-level module syntax
