+++ book/src/SUMMARY.md
@@ -5,9 +5,9 @@
 - [Installation](./installation/index.md)
 - [Language](./language/index.md)
   - [Conditionals](./language/conditionals.md)
-  - [Records](./language/records.md)
   - [Functions](./language/functions.md)
-  - [Modules](./language/modules.md)
+  - [Records](./language/records.md)
+  - [Bindings](./language/bindings.md)
   - [Type inference](./language/type-inference.md)
   - [Universes](./language/universes.md)
 - [Appendix](./appendix/index.md)
