+++ src/cli/check.rs
@@ -27,7 +27,7 @@ pub fn run(color: ColorChoice, opts: Opts) -> Result<(), Error> {
     let mut is_error = false;
     for path in opts.files {
         let file = codemap.add_filemap_from_disk(path)?;
-        let (concrete_module, _import_paths, parse_errors) = parse::module(&file);
+        let (concrete_term, _import_paths, parse_errors) = parse::term(&file);
 
         let mut is_parse_error = false;
         for error in parse_errors {
@@ -39,8 +39,8 @@ pub fn run(color: ColorChoice, opts: Opts) -> Result<(), Error> {
             continue;
         }
 
-        let raw_module = match concrete_module.desugar(&desugar_env) {
-            Ok(raw_module) => raw_module,
+        let raw_term = match concrete_term.desugar(&desugar_env) {
+            Ok(raw_term) => raw_term,
             Err(err) => {
                 codespan_reporting::emit(&mut writer.lock(), &codemap, &err.to_diagnostic())?;
                 is_error = true;
@@ -48,7 +48,7 @@ pub fn run(color: ColorChoice, opts: Opts) -> Result<(), Error> {
             },
         };
 
-        match semantics::check_module(&tc_env, &raw_module) {
+        match semantics::infer_term(&tc_env, &raw_term) {
             Ok(_) => {},
             Err(err) => {
                 codespan_reporting::emit(&mut writer.lock(), &codemap, &err.to_diagnostic())?;
