+++ src/semantics/tests/mod.rs
@@ -8,21 +8,6 @@ use syntax::translation::{Desugar, DesugarEnv};
 
 use super::*;
 
-fn parse_module(codemap: &mut CodeMap, src: &str) -> concrete::Module {
-    let filemap = codemap.add_filemap(FileName::virtual_("test"), src.into());
-    let (concrete_module, _import_paths, errors) = parse::module(&filemap);
-
-    if !errors.is_empty() {
-        let writer = StandardStream::stdout(ColorChoice::Always);
-        for error in errors {
-            codespan_reporting::emit(&mut writer.lock(), &codemap, &error.to_diagnostic()).unwrap();
-        }
-        panic!("parse error!")
-    }
-
-    concrete_module
-}
-
 fn parse_term(codemap: &mut CodeMap, src: &str) -> concrete::Term {
     let filemap = codemap.add_filemap(FileName::virtual_("test"), src.into());
     let (concrete_term, _import_paths, errors) = parse::term(&filemap);
@@ -78,7 +63,6 @@ fn parse_check_term(codemap: &mut CodeMap, tc_env: &TcEnv, src: &str, expected: 
     }
 }
 
-mod check_module;
 mod check_term;
 mod infer_term;
 mod normalize;
